Templatization Utilities

Impostor._materialize_template — Function

_materialize_template(template::String; locale::String) :: String
_materialize_template(template::String, reference_dfrow::DataFrames.DataFrameRow; locale::String) :: String

Impostor._materialize_numeric_template — Function

_materialize_numeric_template(template::String) :: String
_materialize_numeric_template(template::AbstractString, number::Integer) :: String

Receive a numeric template string (e.g. "###-#") and generate a string replacing the '#' chars by random integers between [0, 9]; pass fixed numbers in the numeric template (e.g. "(15) 9####-####") to pre-select the numbers in the returned string.

Optionally, provide a number to fill the placeholders '#' in template. In this usage, the number of digits in number may be greater of equal to the number of '#' in template, but not smaller.

Examples

julia> Impostor._materialize_numeric_template("####-#")
"1324-8"

julia> Impostor._materialize_numeric_template("1/####-9")
"1/5383-9"

julia> Impostor._materialize_numeric_template("####-#", 12345)
"1234-5"

julia> Impostor._materialize_numeric_template("1/####-9", 4321)
"1/4321-9"

julia> Impostor._materialize_numeric_template("####", 87654321)
"8765"

Impostor._materialize_numeric_range_template — Function

_materialize_numeric_range_template(template::AbstractString) :: String

Generate a string containing a number from a numeric range template. Such numeric templates may contain options separated by a ';' caracter. Additionally, options can assume a single template format (e.g. "4##") or specify a range using the ':' character inside the option (e.g. "2##:3##" specifies numbers between 200 and 399).

Example

julia> Impostor._materialize_numeric_range_template("4#####")
"412345"

julia> Impostor._materialize_numeric_range_template("34####;37####")  # will select 34#### or 37####
"349790"

julia> Impostor._materialize_numeric_range_template("51####:55####")
"532489"

julia> Impostor._materialize_numeric_range_template("2221##:2720##;51####:55####")  # will select 2221##:2720## or 51####:55####
"250000"