Data Interface

Impostor._load! — Function

_load!(provider::T, content::T, locale::Vector{T}) :: DataFrame where {T <: AbstractString}
_load!(provider::T, content::T, locale::T = "noloc") :: DataFrame where {T <: AbstractString}

Fetch from the data archive the content associated to the provided locale and provider. Data is returned as a DataFrame for further manipulation. Optionally provide locale = "noloc for the specific contents without any locale assocated to them.

Parameters

provider::AbstractString: provider name, e.g. "localization".

content::AbstractString: content name, e.g. "street_prefix".

locale::Union{AbstractString, Vector{AbstractString}} = "nolod": locale(s) associated to the content and provider provided. Defaults to the "noloc" placeholder for contents which are considered "locale-less".

Impostor._provider_loaded — Function

_provider_loaded(d::DataContainer, provider::AbstractString) :: Bool

Return whether the DataContainer d has already loaded the information associated to the content c.

Impostor._content_loaded — Function

_content_loaded(d::DataContainer, provider::T, content::T) :: Bool

Return whether the DataContainer d has already loaded the information associated to the content c from provider p.

Impostor._locale_loaded — Function

_locale_loaded(d::DataContainer, provider::T, content::T, locale::T) :: Bool

Return whether the DataContainer d has already loaded the information associated to the content c from provider p related to locale l.