Impostor Templates

Impostor.ImpostorTemplate — Type

ImpostorTemplate(format::Vector{Symbol})

Struct storing the formats used to

Impostor.ImpostorTemplate — Type

(impostor::ImpostorTemplate)(n::Integer = 1, sink = Dict; kwargs...)

Generates n entries with information specified in the formats.

Parameters

n: number of entries/rows to generate in each format

sink = Dict: type sink for the generated table.

Kwargs

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Examples

Impostor._verify_formats — Method

_sanitize_formats(formats::Vector)

Verify if all formats are available and exported by Impostor.jl, otherwise throw