Utilities

The following utility functions are available:

Impostor.coerse_string_type — Function

coerse_string_type(v::Vector{<:AbstractString}) :: Union{String, Vector{String}}

Automatically unpack the return value of a generator function into a single string, when appropriate.

Parameters

v::Vector{<:AbstractString}: value(s) returned by some generator function.

Example

julia> Impostor.coerse_string_type(["Mark"])
"Mark"

julia> Impostor.coerse_string_type(["Mark", "Jane"])
2-element Vector{String}:
 "Mark"
 "Jane"

Impostor.setlocale! — Function

setlocale!(loc::String)
setlocale!(loc::Vector{<:AbstractString})

Set loc as the default locale for the current session.

Impostor.session_locale — Function

session_locale() :: Vector{String}

Return the current session locale.

Impostor.resetlocale! — Function

resetlocale!()

Reset the current session locale to "en_US".

Impostor.provider_exists — Function

provider_exists(p::AbstractString) :: Bool

Return whether the provided p is available.

Parameters

p::AbstractString: provider name

Impostor.content_exists — Function

content_exists(p::T, c::T) :: Bool where {T <: AbstractString}

Return whether the content c is available for provider p.

Parameters

p::AbstractString: provider name

c::AbstractString: content name

Impostor.locale_exists — Function

locale_exists(p::T, c::T, l::T) :: Bool where {T <: AbstractString}

Return whether the provided locale l is available for content c from provider p.

Parameters

p::AbstractString: provider name

c::AbstractString: content name

l::AbstractString: locale name