Finance

The following generator functions are available in the Finance provider:

Impostor.bank_name — Function

bank_name(n::Integer = 1; kwargs...)
bank_name(options::Vector, n::Integer; level::Symbol, kwargs...)
bank_name(mask::Vector; level::Symbol, kwargs...)

Parameters

n::Integer = 1: number of bank name entries to generate

options::Vector{<:AbstractString}: vector with options restricting the possible values generated.

mask::Vector{<:AbstractString}: mask vector with element-wise option restrictions.

Kwargs

level::Symbol = :bank_code: Level of values in options or mask when using option-based or mask-based eneration.

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Example

julia> bank_name(5; locale = ["pt_BR"])
5-element Vector{String}:
 "Broker"
 "Nubank"
 "Itaubank"
 "Renascenca"
 "Daycoval"

Impostor.bank_official_name — Function

bank_official_name(n::Integer = 1; kwargs...)
bank_official_name(options::Vector, n::Integer; level::Symbol, kwargs...)
bank_official_name(mask::Vector; level::Symbol, kwargs...)

Parameters

n::Integer = 1: number of official bank name entries to generate

options::Vector{<:AbstractString}: vector with options restricting the possible values generated.

mask::Vector{<:AbstractString}: mask vector with element-wise option restrictions.

Kwargs

level::Symbol = :bank_code: Level of values in options or mask when using option-based or mask-based eneration.

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Example

```jldoctest

Impostor.credit_card_cvv — Function

credit_card_cvv(n::Integer = 1; kwargs...)

Generate n credit card ccvs, e.g. 034

Impostor.credit_card_expiry — Function

credit_card_expiry(n::Integer = 1; kwargs...)

Generate n credit card expiry entries, e.g. 05/2029.

Impostor.credit_card_number — Function

credit_card_number(n::Integer = 1; kwargs...)
credit_card_number(options::Vector{<:AbstractString}, n::Integer; kwargs...)
credit_card_number(mask::Vector{<:AbstractString}; kwargs...)

Parameters

n::Integer = 1: number of credit card numbers to generate.

options::Vector{<:AbstractString}: vector with options restricting the possible values generated.

mask::Vector{<:AbstractString}: mask vector with element-wise option restrictions.

Kwargs

formatted::Bool: whether to return the raw credit card numbers e.g. "3756808757861311" or to format the output e.g. "3756-8087-5786-1311"

Impostor.credit_card_vendor — Function

credit_card_vendor(n::Integer = 1; kwargs...)
credit_card_vendor(options::Vector{<:AbstractString}, n::Integer; kwargs...)