Localization

The following generator functions are available in the Localization provider:

Impostor.address — Function

address(n::Integer = 1; kwargs...)
address(options::Vector{<:AbstractString}, n::Integer = 1; level::Symbol, kwargs...)
address(mask::Vector{<:AbstractString}; level::Symbol, kwargs...)

Parameters

n::Integer = 1: number of addresses to generate.

options::Vector{<:AbstractString}: vector with options restricting the possible values generated.

mask::Vector{<:AbstractString}: mask vector with element-wise option restrictions.

Kwargs

level::Symbol = :district_name: option level to be used when using option-based generation. Valid level values are:

:country_code

:state_code

:city_name

:district_name

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.address_complement — Function

address_complement(n::Integer = 1; kwargs...)

Parameters

n::Integer = 1: number of address complements to generate.

Kwargs

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.city — Function

city(n::Integer = 1; kwargs...)
city(options::Vector{<:AbstractString}, n::Integer; level, kwargs...)
city(mask::Vector{<:AbstractString}; level::Symbol, kwargs...)

Parameters

n::Integer = 1: number of country codes to generate.

options::Vector{<:AbstractString}: vector with options restricting the possible values generated.

mask::Vector{<:AbstractString}: mask vector with element-wise option restrictions.

Kwargs

level::Symbol = :city_name: Level of values in options or mask when using option-based or mask-based eneration. Valid level values are:

:country_code

:state_code

:city_name

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.country — Function

country(n::Integer = 1; kwargs...)
country(options::Vector{<:AbstractString}, n::Integer; level::Symbol, kwargs...)
country(mask::Vector{<:AbstractString}; level::Symbol, kwargs...)

Generate n or length(mask) country names.

Parameters

n::Integer = 1: number of country name entries to generate.

options::Vector{<:AbstractString}: vector with with options restricting the possible values generated.

mask::Vector{<:AbstractString}: mask vector with element-wise option restrictions.

Kwargs

level::Symbol = :country_code: Level of values in options or mask when using option-based or mask-based generation. Valid level values are:

:country_code

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.country_code — Function

country_code(n::Integer = 1; kwargs...)
country_code(options::Vector{<:AbstractString}, n::Integer; level::Symbol, kwargs...)
country_code(mask::Vector{<:AbstractString}; level::Symbol, keargs...)

Generate n or length(mask) country codes.

Parameters

n::Integer = 1: number of country codes to generate.

options::Vector{<:AbstractString}: vector with options restricting the possible values generated.

mask::Vector{<:AbstractString}: mask vector with element-wise option restrictions.

Kwargs

level::Symbol = :country_code: Level of values in options or mask when using option-based or mask-based eneration. Valid level values are:

:country_code

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.district — Function

district(n::Integer = 1; kwargs...)
district(options::Vector{<:AbstractString}, n::Integer; level::Symbol, kwargs...)
district(mask::Vector{<:AbstractString}; level::Symbol, kwargs...)

Parameters

n::Integer = 1: number of district names to generate.

options::Vector{<:AbstractString}: vector with options restricting the possible values generated.

mask::Vector{<:AbstractString}: mask vector with element-wise option restrictions.

Kwargs

level::Symbol = :district_name: Level of values in options or mask when using option-based or mask-based eneration. Valid level values are:

:country_code

:state_code

:city_name

:district_name

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.postcode — Function

postcode(n::Integer = 1; kwargs...)

Parameters

n::Integer = 1: number of postcodes to generate.

Kwargs

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.state — Function

state(n::Integer = 1; kwargs...)
state(options::Vector{<:AbstractString}, n::Integer; level::Symbol, kwargs...)
state(mask::Vector{<:AbstractString}; level::Symbol, kwargs...)

Parameters

n::Integer = 1: number of country codes to generate.

options::Vector{<:AbstractString}: vector with options restricting the possible values generated.

mask::Vector{<:AbstractString}: mask vector with element-wise option restrictions.

Kwargs

level::Symbol = :state_code: Level of values in options or mask when using option-based or mask-based eneration. Valid level values are:

:country_code

:state_code

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.state_code — Function

state_code(n::Integer = 1; kwargs...)
state_code(options::Vector{<:AbstractString}, n::Integer; level::Symbol, kwargs...)
state_code(mask::Vector{<:AbstractString}; level::Symbol, kwargs...)

Parameters

n::Integer = 1: number of state codes to generate.

options::Vector{<:AbstractString}: vector with options restricting the possible values generated.

mask::Vector{<:AbstractString}: mask vector with element-wise option restrictions.

Kwargs

level::Symbol = :state_code: Level of values in options or mask when using option-based or mask-based eneration. Valid level values are:

:country_code

:state_code

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.street — Function

street(n::Integer = 1; kwargs...)

Generate n street names.

Kwargs

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.street_suffix — Function

street_suffix(n::Integer = 1; kwargs...)

Kwargs

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.street_prefix — Function

street_prefix(n::Integer = 1; kwargs...)

Kwargs

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.