Localization

The following generator functions are available in the Localization provider:

Impostor.address — Function

address(n::Integer = 1; kws...)
address(options::Vector{<:AbstractString}, n::Integer = 1; level::Symbol, kws...)
address(mask::Vector{<:AbstractString}; level::Symbol, kws...)

Generate n or length(mask) addresses based on options or mask according to the required level. For example, let level = :country_code and options be the required country codes to generate addresses for, e.g. ["BRA", "USA"], the expected output will contain only addresses (in their specific formats) for location in Brazil ("BRA") and the United States ("USA"). Other values may be passed as level.

Parameters

n::Integer = 1: number of addresses to generate.

options::Vector{<:AbstractString}: vector with options restricting the possible values generated.

mask::Vector{<:AbstractString}: mask vector with element-wise option restrictions.

Kwargs

level::Symbol = :state_code: option level to be used when using option and mask-based generation.

Impostor.address_complement — Function

address_complement(n::Integer = 1; kws...)

Parameters

n::Integer = 1: number of address complements to generate.

Kwargs

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.city — Function

city(n::Integer = 1; kws...)
city(options::Vector{<:AbstractString}, n::Integer; level, kws...)
city(mask::Vector{<:AbstractString}; level::Symbol, kws...)

Parameters

n::Integer = 1: number of country codes to generate.

options::Vector{<:AbstractString}: vector with options restricting the possible values generated.

mask::Vector{<:AbstractString}: mask vector with element-wise option restrictions.

Kwargs

level::Symbol = :state_code: Level of values in options or mask when using option-based or mask-based eneration.

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.country — Function

country(n::Integer = 1; kws...)
country(options::Vector{<:AbstractString}, n::Integer; level::Symbol, kws...)
country(mask::Vector{<:AbstractString}; level::Symbol, kws...)

Generate n or length(mask) country names.

Parameters

n::Integer = 1: number of country name entries to generate.

options::Vector{<:AbstractString}: vector with with options restricting the possible values generated.

mask::Vector{<:AbstractString}: mask vector with element-wise option restrictions.

Kwargs

level::Symbol = :country_code: Level of values in options or mask when using option-based or mask-based generation.

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.country_code — Function

country_code(n::Integer = 1; kws...)
country_code(options::Vector{<:AbstractString}, n::Integer; level::Symbol, kws...)
country_code(mask::Vector{<:AbstractString}; level::Symbol, kws...)

Generate n or length(mask) country codes.

Parameters

n::Integer = 1: number of country codes to generate.

options::Vector{<:AbstractString}: vector with options restricting the possible values generated.

mask::Vector{<:AbstractString}: mask vector with element-wise option restrictions.

Kwargs

level::Symbol = :country_code: Level of values in options or mask when using option-based or mask-based eneration.

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.country_official_name — Function

country_official_name(n::Integer = 1; kws...)
country_official_name(options::Vector{<:AbstractString}, n::Integer; level::Symbol, kws...)
country_official_name(mask::Vector{<:AbstractString}; level::Symbol, kws...)

Generate n or length(mask) country offiical names.

Parameters

n::Integer = 1: number of country official names entries to generate.

options::Vector{<:AbstractString}: vector with with options restricting the possible values generated.

mask::Vector{<:AbstractString}: mask vector with element-wise option restrictions.

Kwargs

level::Symbol = :country_code: Level of values in options or mask when using option-based or mask-based generation.

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.district — Function

district(n::Integer = 1; kws...)
district(options::Vector{<:AbstractString}, n::Integer; level::Symbol, kws...)
district(mask::Vector{<:AbstractString}; level::Symbol, kws...)

Parameters

n::Integer = 1: number of district names to generate.

options::Vector{<:AbstractString}: vector with options restricting the possible values generated.

mask::Vector{<:AbstractString}: mask vector with element-wise option restrictions.

Kwargs

level::Symbol = :district_name: Level of values in options or mask when using option-based or mask-based eneration.

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.postcode — Function

postcode(n::Integer = 1; kws...)

Parameters

n::Integer = 1: number of postcodes to generate.

Kwargs

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.state — Function

state(n::Integer = 1; kws...)
state(options::Vector{<:AbstractString}, n::Integer; level::Symbol, kws...)
state(mask::Vector{<:AbstractString}; level::Symbol, kws...)

Parameters

n::Integer = 1: number of country codes to generate.

options::Vector{<:AbstractString}: vector with options restricting the possible values generated.

mask::Vector{<:AbstractString}: mask vector with element-wise option restrictions.

Kwargs

level::Symbol = :country_code: Level of values in options or mask when using option-based or mask-based eneration.

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.state_code — Function

state_code(n::Integer = 1; kws...)
state_code(options::Vector{<:AbstractString}, n::Integer; level::Symbol, kws...)
state_code(mask::Vector{<:AbstractString}; level::Symbol, kws...)

Parameters

n::Integer = 1: number of state codes to generate.

options::Vector{<:AbstractString}: vector with options restricting the possible values generated.

mask::Vector{<:AbstractString}: mask vector with element-wise option restrictions.

Kwargs

level::Symbol = :country_code: Level of values in options or mask when using option-based or mask-based eneration.

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.street — Function

street(n::Integer = 1; kws...)

Generate n street names.

Kwargs

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.street_number — Function

street_number(n::Integer = 1)

Parameters

n::Integer = 1: number of street numbers to generate.

Impostor.street_prefix — Function

street_prefix(n::Integer = 1; kws...)

Kwargs

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.

Impostor.street_suffix — Function

street_suffix(n::Integer = 1; kws...)

Kwargs

locale::Vector{String}: locale(s) from which entries are sampled. If no locale is provided, the current session locale is used.